Application から Controllerを分離 #2
